Move django-csrf.js and rtd-import.js into core's static-src directory
